Debugging skeletal animation.
It seems that .sch3Darmature parsing is working, but the bone weight
parsing in .sch3Dmesh is not.

This references #2.

+++ Schooner 3D/src/com/supermercerbros/gameengine/material/CelShadedMaterial.java
@@ -83,7 +83,7 @@ public class CelShadedMaterial extends Material {
 	}
 	
 	@Override
-	public void onLoadObject(GameObject obj, int[] vbo, int vertCount) {
+	public void onLoadObject(GameObject obj, float[] vbo, int vertCount) {
 		loadArrayToVbo(obj.verts, vbo, 3, vertCount);
 		loadArrayToVbo(obj.normals, vbo, 3, vertCount);
 		loadArrayToVbo(obj.mtl, vbo, 2, vertCount);
